model 이름 및 Mapper 이름 수정
