+++ core/data/src/main/java/com/depromeet/team5/core/data/mapper/DataMapper.kt
@@ -1,7 +1,7 @@
 package com.depromeet.team5.core.data.mapper
 
 
-interface DomainMapper<out T> {
+interface DataMapper<out T> {
 
     fun toDomain(): T
 }
