+++ core/remotedatasource/src/main/java/com/depromeet/team5/core/remotedatasource/apisource/HedgeApiSource.kt
@@ -1,9 +1,9 @@
 package com.depromeet.team5.core.remotedatasource.apisource
 
-import com.depromeet.team5.core.remotedatasource.model.RetrospectionResponse
+import com.depromeet.team5.core.remotedatasource.model.RetrospectionRemoteData
 
 
 interface HedgeApiSource {
 
-    suspend fun createRetrospection(body: Map<String, Any?>): RetrospectionResponse
+    suspend fun createRetrospection(body: Map<String, Any?>): RetrospectionRemoteData
 }
