+++ core/retrofit/src/main/java/com/depromeet/team5/core/retrofit/api/HedgeApi.kt
@@ -1,6 +1,6 @@
 package com.depromeet.team5.core.retrofit.api
 
-import com.depromeet.team5.core.retrofit.model.RetrospectionDto
+import com.depromeet.team5.core.retrofit.model.RetrospectionResponse
 import okhttp3.RequestBody
 import retrofit2.http.Body
 import retrofit2.http.POST
@@ -9,5 +9,5 @@ import retrofit2.http.POST
 interface HedgeApi {
 
     @POST("api/v1/retrospections")
-    suspend fun createRetrospection(@Body body: RequestBody): RetrospectionDto
+    suspend fun createRetrospection(@Body body: RequestBody): RetrospectionResponse
 }
