+++ core/retrofit/src/main/java/com/depromeet/team5/core/retrofit/apisourceimpl/HedgeApiSourceImpl.kt
@@ -14,5 +14,5 @@ internal class HedgeApiSourceImpl @Inject constructor(
 
     override suspend fun createRetrospection(
         body: Map<String, Any?>
-    ) = hedgeApi.createRetrospection(body.toRequestBody()).toRemoteResponse()
+    ) = hedgeApi.createRetrospection(body.toRequestBody()).toRemoteData()
 }
