--- core/retrofit/src/main/java/com/depromeet/team5/core/retrofit/mapper/ResponseMapper.kt
@@ -1,7 +1,0 @@
-package com.depromeet.team5.core.retrofit.mapper
-
-
-internal interface ResponseMapper<out T> {
-
-    fun toRemoteResponse(): T
-}
