+++ core/retrofit/src/main/java/com/depromeet/team5/core/retrofit/mapper/RetrofitMapper.kt
@@ -1,0 +1,7 @@
+package com.depromeet.team5.core.retrofit.mapper
+
+
+internal interface RetrofitMapper<out T> {
+
+    fun toRemoteData(): T
+}
